pyproject.toml
[project]
name = "project-url-shortener"
version = "0.1.0"
description = ""
authors = [
    {name = "rubenadriel",email = "ruben.a.angelo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=45.0.6,<46.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib (==1.7.4)",
    "bcrypt (==4.0.1)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
